move deprecated code to files
